requirements.txt
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
cs50==9.3.0
Flask==3.0.0
Flask-Session==0.5.0
greenlet==3.0.1
idna==3.4
igdb-api-v4==0.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
protobuf==4.24.4
requests==2.31.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
termcolor==2.3.0
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
wheel==0.41.3
